pyproject.toml
[project]
name = "notes2latex"
version = "0.1.0"
description = "Convert handwritten math notes to compiled LaTeX using vision LLMs"
readme = "README.md"
authors = [
    { name = "advaypakhale", email = "advay.pakhale@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "langgraph>=1.0.9",
    "litellm>=1.81.15",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "rich>=14.3.3",
    "tenacity>=9.1.4",
    "typer>=0.24.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.21.0",
    "python-multipart>=0.0.20",
    "sse-starlette>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/api", "src/cli", "src/clients", "src/compiler", "src/core", "src/db", "src/document"]

[tool.hatch.build.targets.wheel.force-include]
"src/app.py" = "app.py"

[project.scripts]
notes2latex = "cli.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
